move table child components in their own file
